feat: add login as admin even if login is disabled
