+++ src/controllers/admin/auth/login.ts
@@ -1,0 +1,11 @@
+import { NextFunction, Request, Response } from 'express';
+import { login } from '../../../utils/login';
+
+export default [
+  // Middlewares
+
+  // Controller
+  async (request: Request, response: Response, next: NextFunction) => {
+    await login(request, response, next, true);
+  },
+];
